requirements.txt
Django>=4.2.0,<5.0.0
Pillow>=9.0.0
django-crispy-forms>=2.0
crispy-bootstrap5>=0.7